ai-service/requirements.txt
flask~=2.2.2
opencv-python~=4.6.0.66
numpy~=1.23.3
pandas~=1.3.5
shapely~=2.0.0
matplotlib~=3.6.1
boto3~=1.24.28
pillow~=9.2.0
python-dotenv~=0.19.2
requests~=2.28.2
geopandas~=0.12.2
scikit-learn~=1.0.2
segmentation_models_pytorch
joblib~=1.2.0
scikit-image